fix: CDC review fixes — use i32 for norm_level, validate chunk sizes, clean up proto handling
- Change `CdcOptions::norm_level` from `i64` to `i32` to match parquet's type,
  replacing `config_field!(i64)` with `config_field!(i32)` and removing all
  now-unnecessary `as i32`/`as i64` casts
- Add validation in `into_writer_properties_builder` for invalid CDC config:
  `min_chunk_size == 0` and `max_chunk_size <= min_chunk_size`, returning a
  proper `DataFusionError::Configuration` instead of panicking in the parquet layer
- Add explanatory comments on why proto zero-value handling is asymmetric
  between chunk sizes (0 is invalid) and `norm_level` (0 is valid default)
- Remove extra blank line in `ParquetOptions` config block
- Remove unused `parquet` feature from `datafusion-proto-common/Cargo.toml`
- Add three validation tests for the new error paths

+++ datafusion/common/src/config.rs
@@ -701,7 +701,7 @@ config_namespace! {
 
         /// Normalization level. Increasing this improves deduplication ratio
         /// but increases fragmentation. Recommended range is [-3, 3], default is 0.
-        pub norm_level: i64, default = 0
+        pub norm_level: i32, default = 0
     }
 }
 
@@ -862,7 +862,6 @@ config_namespace! {
         /// (writing) Sets bloom filter number of distinct values. If NULL, uses
         /// default parquet writer setting
         pub bloom_filter_ndv: Option<u64>, default = None
-
 
         /// (writing) Controls whether DataFusion will attempt to speed up writing
         /// parquet files by serializing them in parallel. Each column
@@ -1851,7 +1850,7 @@ config_field!(usize);
 config_field!(f64);
 config_field!(u64);
 config_field!(u32);
-config_field!(i64);
+config_field!(i32);
 
 impl ConfigField for u8 {
     fn visit<V: Visit>(&self, v: &mut V, key: &str, description: &'static str) {
